Delete original server.js file and rename app.js file to server.js
